Using retrofit to connect borrower service to loan service

+++ src/main/kotlin/net/mreunionlabs/services/loan/Model.kt
@@ -1,3 +1,3 @@
 package net.mreunionlabs.services.loan
 
-data class Loan(val id:Int, val amount: Int, val borrowerId: Int)
+data class Loan(val id:Long, val amount: Int, val borrowerId: Int)
